requirements.txt
fastapi                    ==  0.115.4
numpy                      ==  1.26.1
uvicorn[standard]          ==  0.32.0
httpx                      ==  0.27.2
psycopg2-binary            ==  2.9.10
pyjwt                      ==  2.10.1
requests                   ==  2.32.3
together                   ==  1.4.6
